requirements.txt
ebooklib
pyyaml
click
schema
requests
beautifulsoup4
lxml
python-box[all]>=5.0,<6.0